+++ CHANGELOG.md
@@ -9,6 +9,8 @@
    | [#420](https://github.com/JetBrains/bazel-bsp/pull/420)
 -  Fix ide classpath computation for recent rules_jvm_external.
    | [#421](https://github.com/JetBrains/bazel-bsp/pull/421)
+-  Process exit instead of hang in case of uncaught exception in pooled threads
+   | [#425](https://github.com/JetBrains/bazel-bsp/pull/425)
 
 ## [2.7.0]